requirements.txt
llama-cpp-python
llama-index-core
llama-index-embeddings-huggingface